requirements.txt
feedparser
feedgen
PyYAML